Don't delete a species if already deleted during the teardown phase, otherwise you get errors.

+++ controller/generate-controller.js
@@ -99,10 +99,18 @@ async function tearDownProduct(namespace) {
 async function tearDownSpecies(namespace) {
     log.info(`${CONTROLLER_NAME}::tearDownSpecies:${namespace}`)
     const speciesList = localStorage.getItem(namespace, 'speciesList')
+    var deleted = []
     if (speciesList) {
         for (productNo of speciesList) {
-            log.info(`${CONTROLLER_NAME}::about to teardown species for product with product no. ${productNo}`)
-            await speciesService.deleteSpecies(productNo)
+            if (deleted.indexOf(productNo) == -1) {
+                log.info(`${CONTROLLER_NAME}::about to teardown all species for product with product no. ${productNo}`)
+                // Will delete ALL species for a specific product.
+                await speciesService.deleteSpecies(productNo)
+                // Track the species we have deleted for each product, in case the test data contained
+                // multiple species for the same product. Otherwise, we will try and delete species for a
+                // product that have already been deleted, causing an error.
+                deleted.push(productNo)
+            }
         }
     }
 }
